[GH-#274] changed naming folder structure
